+++ Maple-Building-Management-App/Views/Complaint/ComplaintDetails.cshtml
@@ -1,0 +1,65 @@
+﻿@model Maple_Building_Management_App.Models.ComplaintModel
+
+@{
+    ViewBag.Title = "Complaint Details";
+}
+
+@using Maple_Building_Management_App.Models;
+
+<h2>Complaint Details</h2>
+
+
+@using (Html.BeginForm())
+{
+<div class="form-horizontal">
+    <h4></h4>
+    <hr />
+    <div class="form-group">
+        @Html.LabelFor(model => model.ComplaintStatus, htmlAttributes: new { @class = "control-label col-md-2" })
+        <div class="col-md-10">
+            @Html.DisplayFor(model => model.ComplaintStatus)
+        </div>
+    </div>
+
+    <div class="form-group">
+        @Html.LabelFor(model => model.ComplaintType, htmlAttributes: new { @class = "control-label col-md-2" })
+        <div class="col-md-10">
+            @Html.DisplayFor(model => model.ComplaintType)
+        </div>
+    </div>
+
+    <div class="form-group">
+        @Html.LabelFor(model => model.Description, htmlAttributes: new { @class = "control-label col-md-2" })
+        <div class="col-md-10">
+            @Html.DisplayFor(model => model.Description)
+        </div>
+    </div>
+
+    <div class="form-group">
+        @Html.LabelFor(model => model.IncidentDate, htmlAttributes: new { @class = "control-label col-md-2" })
+        <div class="col-md-10">
+            @Html.DisplayFor(model => model.IncidentDate)
+        </div>
+    </div>
+
+    @{
+        if (Model.ComplaintStatus == "Resolved" && Model.Feedback != null && Model.Feedback != "")
+        {
+            <div class="form-group">
+                @Html.LabelFor(model => model.Feedback, htmlAttributes: new { @class = "control-label col-md-2" })
+                <div class="col-md-10">
+                    @Html.DisplayFor(model => model.Feedback)
+                </div>
+            </div>
+        }
+    }
+</div>
+}
+
+<div>
+    @Html.ActionLink("Back to List", "ViewComplaints")
+</div>
+
+@section Scripts {
+    @Scripts.Render("~/bundles/jqueryval")
+}
