+++ Maple-Building-Management-App/Views/Complaint/EditComplaint.cshtml
@@ -56,12 +56,25 @@
         </div>
     </div>
 
+    @{
+        if (Model.ComplaintStatus == "Resolved")
+        {
+            <div class="form-group">
+                @Html.LabelFor(model => model.Feedback, htmlAttributes: new { @class = "control-label col-md-2" })
+                <div class="col-md-10">
+                    @Html.EditorFor(model => model.Feedback, new { htmlAttributes = new { @class = "form-control" } })
+                    @Html.ValidationMessageFor(model => model.Feedback, "", new { @class = "text-danger" })
+                </div>
+            </div>
+        }
+    }
+
     <div class="form-group">
         <div class="col-md-offset-2 col-md-10">
             <input type="submit" value="Edit" class="btn btn-default" />
             @*<span class="text-danger">
-                @ViewBag.ErrorMessage
-            </span>*@
+                    @ViewBag.ErrorMessage
+                </span>*@
         </div>
     </div>
 </div>
