--- Maple-Building-Management-App/Views/Home/UpdateComplaint.cshtml
@@ -1,73 +1,0 @@
-﻿@model Maple_Building_Management_App.Models.ComplaintModel
-
-@{
-    ViewBag.Title = "Update Complaint";
-}
-
-@using Maple_Building_Management_App.Models;
-
-<h2>Update Complaint</h2>
-
-
-@using (Html.BeginForm()) 
-{
-    @Html.AntiForgeryToken()
-    
-    <div class="form-horizontal">
-        <h4>Update a Complaint</h4>
-        <hr />
-        @Html.ValidationSummary(true, "", new { @class = "text-danger" })   
-
-        <div class="form-group">
-            @Html.LabelFor(model => model.ComplaintStatus, htmlAttributes: new { @class = "control-label col-md-2" })
-            <div class="col-md-10">
-                @Html.DropDownListFor(model => model.ComplaintStatus,
-                    new SelectList(Enum.GetValues(typeof(ComplaintStatus))),
-                    "",
-                    htmlAttributes: new { @class = "form-control" } )
-                @Html.ValidationMessageFor(model => model.ComplaintStatus, "", new { @class = "text-danger" })
-            </div>
-        </div>
-
-        <div class="form-group">
-            @Html.LabelFor(model => model.ComplaintType, htmlAttributes: new { @class = "control-label col-md-2" })
-            <div class="col-md-10">
-                @Html.DropDownListFor(model => model.ComplaintType,
-                    new SelectList(Enum.GetValues(typeof(ComplaintType))),
-                    "",
-                    htmlAttributes: new { @class = "form-control" } )
-                @Html.ValidationMessageFor(model => model.ComplaintType, "", new { @class = "text-danger" })
-            </div>
-        </div>
-
-        <div class="form-group">
-            @Html.LabelFor(model => model.Description, htmlAttributes: new { @class = "control-label col-md-2" })
-            <div class="col-md-10">
-                @Html.EditorFor(model => model.Description, new { htmlAttributes = new { @class = "form-control" } })
-                @Html.ValidationMessageFor(model => model.Description, "", new { @class = "text-danger" })
-            </div>
-        </div>
-
-        <div class="form-group">
-            @Html.LabelFor(model => model.IncidentDate, htmlAttributes: new { @class = "control-label col-md-2" })
-            <div class="col-md-10">
-                @Html.EditorFor(model => model.IncidentDate, new { htmlAttributes = new { @class = "form-control" } })
-                @Html.ValidationMessageFor(model => model.IncidentDate, "", new { @class = "text-danger" })
-            </div>
-        </div>
-
-        <div class="form-group">
-            <div class="col-md-offset-2 col-md-10">
-                <input type="submit" value="Create" class="btn btn-default" />
-            </div>
-        </div>
-    </div>
-}
-
-<div>
-    @Html.ActionLink("Back to List", "Index")
-</div>
-
-@section Scripts {
-    @Scripts.Render("~/bundles/jqueryval")
-}
