+++ Maple-Building-Management-App/Views/Shared/_Layout.cshtml
@@ -35,7 +35,8 @@
                     {
                         <li>@Html.ActionLink("Manage", "Index", "Manage")</li>
                         <li>@Html.ActionLink("View Accounts", "ViewAccounts", "Home")</li>
-                        <li>@Html.ActionLink("File Complaint", "FileComplaint", "Home")</li>
+                        <li>@Html.ActionLink("File Complaint", "FileComplaint", "Complaint")</li>
+                        <li>@Html.ActionLink("View My Complaints", "ViewComplaints", "Complaint")</li>
                         <li>@Html.ActionLink("Logout", "Logout", "Home")</li>
 
                     }
